requirements.txt
streamlit==1.38.0
pypdf==5.0.0
python-dotenv==1.0.1

langchain==0.2.15
langchain-community==0.2.13
langchain-text-splitters==0.2.2

sentence-transformers==2.2.2
huggingface-hub==0.24.6

faiss-cpu==1.8.0.post1
numpy<2

groq==0.12.0
openai==1.51.0
httpx==0.27.2
